Added matching for sell offers from Workers

+++ src/main/java/pl/edu/mimuw/bajttrade/oferty/Rachunek.java
@@ -5,10 +5,10 @@ import pl.edu.mimuw.bajttrade.przedmioty.Przedmiot;
 public class Rachunek {
   private final int dzien;
   private final int ilosc;
-  private final int cena;
+  private final double cena;
   private final Przedmiot przedmiot;
 
-  public Rachunek(int dzien, int ilosc, int cena, Przedmiot przedmiot) {
+  public Rachunek(int dzien, int ilosc, double cena, Przedmiot przedmiot) {
     this.dzien = dzien;
     this.ilosc = ilosc;
     this.cena = cena;
@@ -23,7 +23,7 @@ public class Rachunek {
     return ilosc;
   }
 
-  public int getCena() {
+  public double getCena() {
     return cena;
   }
